src/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Mail==0.9.1
flask-marshmallow==1.2.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.3.1
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
limits==3.10.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.28
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
ulid==1.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
